requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
bcrypt==4.0.1
paho-mqtt==1.6.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
resend==0.6.0
